# public/swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "WotsApp API",
    "contact": {
      "email": "urip.subekti13@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/messages": {
      "get": {
        "tags": ["messages"],
        "summary": "Get messages",
        "description": "Get messages",
        "operationId": "getMessages",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Number of page",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "search query",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Messages"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Messages": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },

          "is_next": {
            "type": "boolean"
          },

          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "session_id": {
            "type": "string"
          },
          "wa_id": {
            "type": "string"
          },
          "jid": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status_code": {
            "type": "integer",
            "format": "int64"
          },
          "status_message": {
            "type": "string"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      
    }
  }
}